+++ packages/acceptance-tests/pkg-tests-specs/sources/commands/add.test.ts
@@ -324,7 +324,7 @@ describe(`Commands`, () => {
         expect(lock).toMatchObject({
           [`inject-node-gyp@npm:^1.0.0`]: {
             dependencies: {
-              [`node-gyp`]: `latest`,
+              [`node-gyp`]: `npm:latest`,
             },
           },
         });