+++ packages/plugin-compat/tests/patches.test.ts
@@ -27,7 +27,8 @@ async function getDescriptorCandidates(descriptor: Descriptor) {
     const resolver = configuration.makeResolver();
     const resolveOptions: ResolveOptions = {project, resolver, report: new ThrowReport()};
 
-    const candidates = await resolver.getCandidates(descriptor, new Map(), resolveOptions);
+    const normalizedDescriptor = configuration.normalizeDependency(descriptor);
+    const candidates = await resolver.getCandidates(normalizedDescriptor, {}, resolveOptions);
 
     return candidates;
   });