+++ packages/plugin-link/sources/RawLinkResolver.ts
@@ -32,7 +32,7 @@ export class RawLinkResolver implements Resolver {
   }
 
   getResolutionDependencies(descriptor: Descriptor, opts: MinimalResolveOptions) {
-    return [];
+    return {};
   }
 
   async getCandidates(descriptor: Descriptor, dependencies: unknown, opts: ResolveOptions) {