+++ packages/plugin-npm/sources/NpmRemapResolver.ts
@@ -1,7 +1,7 @@
-import {Descriptor, Locator, MinimalResolveOptions, ResolveOptions, Resolver, DescriptorHash, Package} from '@yarnpkg/core';
-import {structUtils}                                                                                   from '@yarnpkg/core';
+import {Descriptor, Locator, MinimalResolveOptions, ResolveOptions, Resolver, Package} from '@yarnpkg/core';
+import {structUtils}                                                                   from '@yarnpkg/core';
 
-import {PROTOCOL}                                                                                      from './constants';
+import {PROTOCOL}                                                                      from './constants';
 
 export class NpmRemapResolver implements Resolver {
   supportsDescriptor(descriptor: Descriptor, opts: MinimalResolveOptions) {
@@ -29,19 +29,25 @@ export class NpmRemapResolver implements Resolver {
   }
 
   getResolutionDependencies(descriptor: Descriptor, opts: MinimalResolveOptions) {
-    const nextDescriptor = structUtils.parseDescriptor(descriptor.range.slice(PROTOCOL.length), true);
+    const nextDescriptor = opts.project.configuration.normalizeDependency(
+      structUtils.parseDescriptor(descriptor.range.slice(PROTOCOL.length), true),
+    );
 
     return opts.resolver.getResolutionDependencies(nextDescriptor, opts);
   }
 
-  async getCandidates(descriptor: Descriptor, dependencies: Map<DescriptorHash, Package>, opts: ResolveOptions) {
-    const nextDescriptor = structUtils.parseDescriptor(descriptor.range.slice(PROTOCOL.length), true);
+  async getCandidates(descriptor: Descriptor, dependencies: Record<string, Package>, opts: ResolveOptions) {
+    const nextDescriptor = opts.project.configuration.normalizeDependency(
+      structUtils.parseDescriptor(descriptor.range.slice(PROTOCOL.length), true),
+    );
 
     return await opts.resolver.getCandidates(nextDescriptor, dependencies, opts);
   }
 
   async getSatisfying(descriptor: Descriptor, references: Array<string>, opts: ResolveOptions) {
-    const nextDescriptor = structUtils.parseDescriptor(descriptor.range.slice(PROTOCOL.length), true);
+    const nextDescriptor = opts.project.configuration.normalizeDependency(
+      structUtils.parseDescriptor(descriptor.range.slice(PROTOCOL.length), true),
+    );
 
     return opts.resolver.getSatisfying(nextDescriptor, references, opts);
   }