+++ packages/plugin-npm/sources/NpmSemverResolver.ts
@@ -1,12 +1,12 @@
-import {ReportError, MessageName, Resolver, ResolveOptions, MinimalResolveOptions, Manifest, DescriptorHash, Package, miscUtils} from '@yarnpkg/core';
-import {Descriptor, Locator, semverUtils}                                                                                        from '@yarnpkg/core';
-import {LinkType}                                                                                                                from '@yarnpkg/core';
-import {structUtils}                                                                                                             from '@yarnpkg/core';
-import semver                                                                                                                    from 'semver';
+import {ReportError, MessageName, Resolver, ResolveOptions, MinimalResolveOptions, Manifest, Package, miscUtils} from '@yarnpkg/core';
+import {Descriptor, Locator, semverUtils}                                                                        from '@yarnpkg/core';
+import {LinkType}                                                                                                from '@yarnpkg/core';
+import {structUtils}                                                                                             from '@yarnpkg/core';
+import semver                                                                                                    from 'semver';
 
-import {NpmSemverFetcher}                                                                                                        from './NpmSemverFetcher';
-import {PROTOCOL}                                                                                                                from './constants';
-import * as npmHttpUtils                                                                                                         from './npmHttpUtils';
+import {NpmSemverFetcher}                                                                                        from './NpmSemverFetcher';
+import {PROTOCOL}                                                                                                from './constants';
+import * as npmHttpUtils                                                                                         from './npmHttpUtils';
 
 const NODE_GYP_IDENT = structUtils.makeIdent(null, `node-gyp`);
 const NODE_GYP_MATCH = /\b(node-gyp|prebuild-install)\b/;
@@ -39,10 +39,10 @@ export class NpmSemverResolver implements Resolver {
   }
 
   getResolutionDependencies(descriptor: Descriptor, opts: MinimalResolveOptions) {
-    return [];
+    return {};
   }
 
-  async getCandidates(descriptor: Descriptor, dependencies: Map<DescriptorHash, Package>, opts: ResolveOptions) {
+  async getCandidates(descriptor: Descriptor, dependencies: Record<string, Package>, opts: ResolveOptions) {
     const range = semverUtils.validRange(descriptor.range.slice(PROTOCOL.length));
     if (range === null)
       throw new Error(`Expected a valid range, got ${descriptor.range.slice(PROTOCOL.length)}`);
@@ -174,7 +174,7 @@ export class NpmSemverResolver implements Resolver {
 
       conditions: manifest.getConditions(),
 
-      dependencies: manifest.dependencies,
+      dependencies: opts.project.configuration.normalizeDependencyMap(manifest.dependencies),
       peerDependencies: manifest.peerDependencies,
 
       dependenciesMeta: manifest.dependenciesMeta,