+++ packages/plugin-npm/sources/NpmTagResolver.ts
@@ -32,7 +32,7 @@ export class NpmTagResolver implements Resolver {
   }
 
   getResolutionDependencies(descriptor: Descriptor, opts: MinimalResolveOptions) {
-    return [];
+    return {};
   }
 
   async getCandidates(descriptor: Descriptor, dependencies: unknown, opts: ResolveOptions) {