+++ packages/yarnpkg-core/sources/Project.ts
@@ -42,7 +42,7 @@ import {IdentHash, DescriptorHash, LocatorHash, PackageExtensionStatus} from './
 // When upgraded, the lockfile entries have to be resolved again (but the specific
 // versions are still pinned, no worry). Bump it when you change the fields within
 // the Package type; no more no less.
-const LOCKFILE_VERSION = 6;
+const LOCKFILE_VERSION = 7;
 
 // Same thing but must be bumped when the members of the Project class changes (we
 // don't recommend our users to check-in this file, so it's fine to bump it even
@@ -355,7 +355,14 @@ export class Project {
           this.originalPackages.set(pkg.locatorHash, pkg);
 
           for (const entry of key.split(MULTIPLE_KEYS_REGEXP)) {
-            const descriptor = structUtils.parseDescriptor(entry);
+            let descriptor = structUtils.parseDescriptor(entry);
+
+            // Yarn pre-v4 used to generate descriptors that were sometimes
+            // missing the `npm:` protocol
+            if (lockfileVersion <= 6) {
+              descriptor = this.configuration.normalizeDependency(descriptor);
+              descriptor = structUtils.makeDescriptor(descriptor, descriptor.range.replace(/^patch:[^@]+@(?!npm(:|%3A))/, `$1npm%3A`));
+            }
 
             this.storedDescriptors.set(descriptor.descriptorHash, descriptor);
             this.storedResolutions.set(descriptor.descriptorHash, locator.locatorHash);
@@ -793,14 +800,18 @@ export class Project {
           return startDescriptorAliasing(descriptor, this.storedDescriptors.get(alias)!);
 
         const resolutionDependenciesList = resolver.getResolutionDependencies(descriptor, resolveOptions);
-        const resolvedDependencies = new Map(await miscUtils.allSettledSafe(resolutionDependenciesList.map(async dependency => {
-          const bound = resolver.bindDescriptor(dependency, dependencyResolutionLocator, resolveOptions);
+        const resolvedDependencies = Object.fromEntries(
+          await miscUtils.allSettledSafe(
+            Object.entries(resolutionDependenciesList).map(async ([dependencyName, dependency]) => {
+              const bound = resolver.bindDescriptor(dependency, dependencyResolutionLocator, resolveOptions);
 
-          const resolvedPackage = await scheduleDescriptorResolution(bound);
-          resolutionDependencies.add(resolvedPackage.locatorHash);
+              const resolvedPackage = await scheduleDescriptorResolution(bound);
+              resolutionDependencies.add(resolvedPackage.locatorHash);
 
-          return [dependency.descriptorHash, resolvedPackage] as const;
-        })));
+              return [dependencyName, resolvedPackage] as const;
+            }),
+          ),
+        );
 
         const candidateResolutions = await miscUtils.prettifyAsyncErrors(async () => {
           return await resolver.getCandidates(descriptor, resolvedDependencies, resolveOptions);